requirements.txt
fastapi==0.75.2
pydantic==1.9.0
yandexcloud==0.158.0
pyjwt==2.3.0
requests==2.27.1
uvicorn==0.17.6
python-dotenv==0.20.0